small refactor; added condition to decide if a message should be included in cross post checking
