+++ Sources/Camille/CrossPost/Sequence+MessageDecorator.swift
@@ -1,0 +1,24 @@
+import Sugar
+
+extension Sequence where Iterator.Element == MessageDecorator {
+    func addAttachment(with builder: SlackMessageAttachmentBuilder, buttonResponder: SlackInteractiveButtonResponderService, handler: @escaping InteractiveButtonResponseHandler) {
+        let array = Array(self)
+        guard let user = array.first?.sender, let message = array.first else { return }
+        let channels = self
+            .flatMap { $0.target?.channel }
+            .map { "<#\($0.id)>" }
+            .joined(separator: ", ")
+        
+        builder.field(short: true, title: "User", value: "<@\(user.id)>")
+        builder.field(short: true, title: "Channels", value: channels)
+        builder.field(short: false, title: "Message Preview", value: message.text.substring(to: 50))
+        for button in CrossPostButton.all {
+            builder.button(
+                name: button.rawValue,
+                text: button.text,
+                responder: buttonResponder,
+                handler: handler
+            )
+        }
+    }
+}
