+++ Sources/Camille/main.swift
@@ -14,6 +14,9 @@ let bot = try SlackBot(
     services: [
         CrossPostService(config: CrossPostServiceConfig(
             timeSpan: 60 * 2,
+            includeMessage: { message in
+                return message.text.components(separatedBy: " ").count > 5
+            },
             reportingTarget: "admins",
             publicWarning: { channel, user in
                 return try SlackMessage()
